refactor devices
