+++ tests/callbacks/test_early_stopper.py
@@ -10,8 +10,7 @@ from torchmetrics import Accuracy
 from composer import Trainer
 from composer.callbacks.early_stopper import EarlyStopper
 from composer.core.time import Time, TimeUnit
-from composer.trainer.devices.device_cpu import DeviceCPU
-from composer.trainer.devices.device_gpu import DeviceGPU
+from composer.devices import DeviceCPU, DeviceGPU
 from tests.common import RandomClassificationDataset, SimpleModel, device
 from tests.metrics import MetricSetterCallback
